% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnAUobs_extract_patids.R
\name{fnAUobs_extract_patids}
\alias{fnAUobs_extract_patids}
\title{Extract the unique patids from an array of AURUM Observation data files}
\usage{
fnAUobs_extract_patids(pa_obs_files)
}
\arguments{
\item{pa_obs_files}{the array of AURUM Extract Observation filenames}
}
\value{
a tibble containing the unique patids in supplied AURUM observations
}
\description{
Extract the unique patids from an array of AURUM Observation data files
}
\examples{
fnAUobs_extract_patids( "Observation_001.txt" )
}
