% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnAUread_tab_obs.R
\name{fnAUread_tab_obs}
\alias{fnAUread_tab_obs}
\title{Read the contents of an AURUM Observation file into a data frame}
\usage{
fnAUread_tab_obs(pDatafile)
}
\arguments{
\item{pDatafile}{the filename of a single AURUM Extract Observation}
}
\value{
a tibble containing the contents of a single AURUM observation file
}
\description{
Read the contents of an AURUM Observation file into a data frame
}
\examples{
fnAUread_tab_obs( "Observation_001.txt" )

}
